Add files via upload
A python script that allows the user to enter two airports and the number of days they would like to travel for. The script will scrape kayak.com and will get the price of the trip for up to a year after a specified start date. For examples,  if you want to take 7 days off and go somewhere, but don't know what departure date would be best money-wise, this script allows you to see when your trip would be cheapest.